Refactor: Movie and tv-show details screens
